Make MaterializedViewTests more stable (#18775)
Make MaterializedViewTests more stable
