+++ fe/fe-core/src/test/java/com/starrocks/planner/MaterializedViewSSBTest.java
@@ -15,17 +15,14 @@
 package com.starrocks.planner;
 
 import com.google.common.collect.Lists;
-import com.starrocks.common.FeConstants;
-import com.starrocks.sql.plan.PlanTestBase;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-public class MaterializedViewSSBTest extends PlanTestBase {
+public class MaterializedViewSSBTest extends MaterializedViewTestBase {
     private static final String MATERIALIZED_DB_NAME = "test_mv";
 
     @BeforeClass
     public static void setUp() throws Exception {
-        FeConstants.USE_MOCK_DICT_MANAGER = true;
         MaterializedViewTestBase.setUp();
 
         starRocksAssert.useDatabase(MATERIALIZED_DB_NAME);
@@ -36,7 +33,6 @@ public class MaterializedViewSSBTest extends PlanTestBase {
 
         // create lineorder_flat_mv
         createMaterializedViews("sql/materialized-view/ssb/", Lists.newArrayList("lineorder_flat_mv"));
-        connectContext.getSessionVariable().setOptimizerExecuteTimeout(300000000);
     }
 
     @Test
